+++ src/main/java/ru/startup/game/chatserver/controller/ChatController.java
@@ -1,6 +1,7 @@
 package ru.startup.game.chatserver.controller;
 
 import lombok.AllArgsConstructor;
+import lombok.extern.java.Log;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.handler.annotation.Payload;
 import org.springframework.messaging.handler.annotation.SendTo;
@@ -15,8 +16,10 @@ import ru.startup.game.chatserver.service.MessageService;
 import ru.startup.game.chatserver.service.UserService;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
+import static ru.startup.game.chatserver.controller.util.AttributeUtil.SESSION_ATTRIBUTE_USERNAME;
+
+@Log
 @RestController
 @AllArgsConstructor
 public class ChatController {
@@ -30,39 +33,26 @@ public class ChatController {
     @MessageMapping("/chat.sendMessage")
     @SendTo("/topic/public")
     public ChatMessageDto sendMessage(@Payload ChatMessageDto chatMessage) {
-        MessageDto message = new MessageDto();
-        UserDto user = userService.findUserByUserName(chatMessage.getSender());
-        message.setUser(user);
-        message.setMessage(chatMessage.getContent());
-        MessageDto saveMessage = messageService.save(message);
-        chatMessage.setId(saveMessage.getId());
-        return chatMessage;
+        return messageService.save(chatMessage);
     }
 
     @GetMapping("chat/history")
     public List<ChatMessageDto> getHistoryMessage(){
-        List<ChatMessageDto> chatMessages;
-        List<MessageDto> messages = chatService.findById(1L).getMessages();
-        chatMessages = messages.stream()
-                .map(messageDto -> new ChatMessageDto(messageDto.getId(), ChatMessageDto.MessageType.CHAT, messageDto.getMessage(), messageDto.getUser().getUserName(),messageDto.getLabels()))
-                .collect(Collectors.toList());
-        return chatMessages;
+        log.info("Send history messages to user");
+        return chatService.getChatHistory();
     }
-
-
 
     @PostMapping("chats")
     public List<ChatDto> getChats(@RequestBody UserDto user){
         UserDto userDto = userService.findUserByUserName(user.getUserName());
-        List<ChatDto> chatsByUserName = chatService.findByUser(userDto);
-        return chatsByUserName;
+        return chatService.findByUser(userDto);
     }
 
     @MessageMapping("/chat.addUser")
     @SendTo("/topic/public")
     public ChatMessageDto addUser(@Payload ChatMessageDto chatMessage,
                                SimpMessageHeaderAccessor headerAccessor) {
-        headerAccessor.getSessionAttributes().put("username", chatMessage.getSender());
+        headerAccessor.getSessionAttributes().put(SESSION_ATTRIBUTE_USERNAME, chatMessage.getSender());
         return chatMessage;
     }
 }
