+++ src/main/java/ru/startup/game/chatserver/controller/util/AttributeUtil.java
@@ -1,0 +1,10 @@
+package ru.startup.game.chatserver.controller.util;
+
+public final class AttributeUtil {
+
+    public static final String SESSION_ATTRIBUTE_USERNAME = "username";
+
+    private AttributeUtil() {
+
+    }
+}
