+++ src/main/java/ru/startup/game/chatserver/model/dto/ChatDto.java
@@ -8,7 +8,6 @@ import lombok.NoArgsConstructor;
 import java.util.List;
 
 @Data
-@Builder
 @AllArgsConstructor
 @NoArgsConstructor
 public class ChatDto {
