+++ src/main/java/ru/startup/game/chatserver/model/dto/LabelDto.java
@@ -1,12 +1,10 @@
 package ru.startup.game.chatserver.model.dto;
 
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
-@Builder
 @AllArgsConstructor
 @NoArgsConstructor
 public class LabelDto {
