+++ src/main/java/ru/startup/game/chatserver/model/dto/MessageDto.java
@@ -1,14 +1,12 @@
 package ru.startup.game.chatserver.model.dto;
 
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.util.List;
 
 @Data
-@Builder
 @AllArgsConstructor
 @NoArgsConstructor
 public class MessageDto {
