+++ src/main/java/ru/startup/game/chatserver/service/LabelService.java
@@ -1,14 +1,32 @@
 package ru.startup.game.chatserver.service;
 
+import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import ru.startup.game.chatserver.model.dto.LabelDto;
 import ru.startup.game.chatserver.model.dto.MessageLabelDto;
+import ru.startup.game.chatserver.model.entity.Label;
+import ru.startup.game.chatserver.model.mapper.LabelMapper;
+import ru.startup.game.chatserver.repository.LabelRepository;
 
 @Service
-public interface LabelService {
+@AllArgsConstructor
+public class LabelService {
 
-   /* public LabelDto findByLabelName(String labelName);
+    private LabelMapper labelMapper;
 
-    public LabelDto save(LabelDto labelDto);*/
+    private LabelRepository labelRepository;
 
-    public void saveMessageLabel(MessageLabelDto messageLabel);
+    private MessageService messageService;
+
+    @Transactional
+    public void saveMessageLabel(MessageLabelDto messageLabel) {
+        Label findLabel = labelRepository.findByLabelName(messageLabel.getLabelName());
+        if( findLabel == null){
+            LabelDto labelDto = new LabelDto();
+            labelDto.setLabelName(messageLabel.getLabelName());
+            findLabel = labelRepository.save(labelMapper.labelDtoToLabel(labelDto));
+        }
+        messageService.updateMessageLabel(messageLabel.getMessageId(), findLabel);
+    }
 }
