+++ src/main/java/ru/startup/game/chatserver/service/UserService.java
@@ -1,14 +1,26 @@
 package ru.startup.game.chatserver.service;
 
+import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
-import ru.startup.game.chatserver.model.dto.ChatDto;
 import ru.startup.game.chatserver.model.dto.UserDto;
+import ru.startup.game.chatserver.model.entity.User;
+import ru.startup.game.chatserver.model.mapper.UserMapper;
+import ru.startup.game.chatserver.repository.UserRepository;
 
-import java.util.List;
+import java.util.Optional;
 
 @Service
-public interface UserService {
+@AllArgsConstructor
+public class UserService {
 
-    public UserDto findUserByUserName(String userName);
+    private UserRepository userRepository;
+
+    private UserMapper userMapper;
+
+    public UserDto findUserByUserName(String userName) {
+        Optional<User> userOptional = userRepository.findUserByUserName(userName);
+        return userMapper.userToUserDto(userOptional.orElse(null));
+    }
+
 
 }
